+++ dto/types.ts
@@ -65,7 +65,16 @@ export type SelectFields<T extends string> =
   | (T | FieldAlias<T> | RawField | string & {})[]
   | { [key in T]?: string } & RawField;
 
-export type OrderByField<T> = { field: T | (string & {}), direction?: 'ASC' | 'DESC' };
+export type OrderByField<T> = {
+  raw?: never,
+  field: T | (string & {}),
+  direction?: 'ASC' | 'DESC'
+} | {
+  raw: string,
+  field?: never,
+  direction?: 'ASC' | 'DESC'
+};
+
 export type GroupByField<T> = T | (string & {}) | T[];
 
 export type ColumnData<T extends string> = Partial<Record<T, any>>;
@@ -103,7 +112,10 @@ export type SoftDeleteOptions = Prettify<{
 
 export type LimitOffset =
   | { limit?: never; offset?: never }
-  | { limit: number; offset?: number };
+  | { limit: Limit; offset?: Offset };
+
+export type Limit = null | number;
+export type Offset = null | number;
 
 export type JoinType = 'INNER' | 'LEFT' | 'RIGHT' | 'FULL';
 
@@ -126,7 +138,7 @@ export type SelectQueryBuilder<ColumnKeys extends string, QueryReturnType> = {
 
 export interface WhereQueryBuilder<ColumnKeys extends string, QueryReturnType> {
   orderBy(fields: OrderByField<ColumnKeys>[]): OrderByQueryBuilder<QueryReturnType>;
-  limit(limit: number): LimitQueryBuilder<QueryReturnType>;
+  limit(limit?: Limit): LimitQueryBuilder<QueryReturnType>;
   groupBy(fields: GroupByField<ColumnKeys>): GroupByQueryBuilder<QueryReturnType>;
   buildQuery(options?: BuildQueryOptions): BuildQueryResult;
   executeQuery<ReturnType = QueryReturnType>(): Promise<ReturnType>;
@@ -138,7 +150,7 @@ export type FromQueryBuilder<ColumnKeys extends string, QueryReturnType> = {
   where(where: WhereCondition<ColumnKeys>): WhereQueryBuilder<ColumnKeys, QueryReturnType>;
   groupBy(fields: GroupByField<ColumnKeys>): GroupByQueryBuilder<QueryReturnType>;
   orderBy(fields: OrderByField<ColumnKeys>[]): OrderByQueryBuilder<QueryReturnType>;
-  limit(limit: number): LimitQueryBuilder<QueryReturnType>;
+  limit(limit?: Limit): LimitQueryBuilder<QueryReturnType>;
   buildQuery(options?: BuildQueryOptions): BuildQueryResult;
   executeQuery<ReturnType = QueryReturnType>(): Promise<ReturnType>;
 }
@@ -147,13 +159,13 @@ export interface JoinQueryBuilder<ColumnKeys extends string, QueryReturnType> ex
 // export interface GroupByQueryBuilder extends FromQueryBuilder { }
 
 export interface OrderByQueryBuilder<T> {
-  limit(limit: number): LimitQueryBuilder<T>;
+  limit(limit?: Limit): LimitQueryBuilder<T>;
   buildQuery(options?: BuildQueryOptions): BuildQueryResult;
   executeQuery<K = T>(): Promise<K>;
 }
 export interface GroupByQueryBuilder<QueryReturnType> extends QueryAction<QueryReturnType> { }
 export interface LimitQueryBuilder<QueryReturnType> {
-  offset(offset: number): OffsetQueryBuilder<QueryReturnType>;
+  offset(offset?: Offset): OffsetQueryBuilder<QueryReturnType>;
   buildQuery(options?: BuildQueryOptions): BuildQueryResult;
   executeQuery<ReturnType = QueryReturnType>(): Promise<ReturnType>;
 }
@@ -183,7 +195,7 @@ export interface InsertQueryBuilder<QueryReturnType> {
 }
 export interface DeleteQueryBuilder<ColumnKeys extends string, QueryReturnType> {
   where(conditions: WhereCondition<ColumnKeys>): WhereQueryBuilder<ColumnKeys, QueryReturnType>;
-  limit(limit: number): LimitQueryBuilder<QueryReturnType>;
+  limit(limit?: Limit): LimitQueryBuilder<QueryReturnType>;
   buildQuery(options?: BuildQueryOptions): BuildQueryResult;
   executeQuery<ReturnType = QueryReturnType>(): Promise<ReturnType>;
 }
@@ -209,7 +221,7 @@ type NotInOperator = { 'NOT_IN'?: any[] };
 type BetweenOperator = { 'BETWEEN'?: [any, any] };
 type NotBetweenOperator = { 'NOT_BETWEEN'?: [any, any] };
 type IsNullOperator = { 'IS_NULL'?: true };
-type IsNotNullOperator = { ['IS_NOT_NULL']?: true };
+type IsNotNullOperator = { 'IS_NOT_NULL'?: true };
 
 type OperatorCondition = Prettify<EqualOperator &
   NotEqualOperator &
@@ -228,12 +240,19 @@ type OperatorCondition = Prettify<EqualOperator &
   IsNotNullOperator
 >;
 
+type RawWhereCondition = {
+  'RAW'?: {
+      sql: string;
+      params?: any[];
+  }
+}
+
 // export type WhereCondition = { [key: string]: OperatorCondition | string | number }
 export type SimpleCondition<ColumnKeys extends string> = {
   [key in ColumnKeys]?: OperatorCondition | string | number | Array<string> | Array<number>;
-} & {
+} & ({
   [key: string]: OperatorCondition | string | number | Array<string> | Array<number>;
-};
+} | RawWhereCondition);
 
 type NonEmptyArray<T> = [T, ...T[]];